users/urls.py
from django.urls import path

from users.views import *
from django.contrib.auth.views import LogoutView

urlpatterns = [
    path('login/', MyLoginView.as_view(), name='login'),
    path('logout/', LogoutView.as_view(next_page='main'), name='logout'),
]
